# Fix performance issue when getting descendants.